Split up library from client bits
