+++ client/src/p9.rs
@@ -1,3 +1,7 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at https://mozilla.org/MPL/2.0/.
+
 use devinfo::{get_devices, DiPropValue};
 use indicatif::{ProgressBar, ProgressStyle};
 use p9ds::proto::{P9Version, Rclunk, Rwrite, Tclunk, Twrite, Version};
