+++ lib/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "softnpu"
+version = "0.2.0"
+edition = "2021"
+
+[dependencies]
+p4rs = { git = "https://github.com/oxidecomputer/p4", branch = "main" }
+serde.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
